clients/raspberry-pi/requirements.txt
picamera2==0.3.17
openai==1.12.0
opencv-python==4.9.0.80
elevenlabs==0.2.27
boto3=1.34.54
python-dotenv=0.21.0